requirements.txt
pyserial
opencv-python
numpy
matplotlib
pillow
google-generativeai